Forward host signals to VM and route supervisor tracing via RPC;
Signals are forwarded via Process.signal() RPC. For SIGINT/SIGQUIT the
supervisor writes the PTY control character (\x03/\x1c) to the master,
which goes through the terminal discipline and works across PID
namespaces. Direct kill() doesn't reach container processes because the
shell runs without job control in a separate PID namespace. Supervisor
tracing events are now forwarded to the CLI through the LogSink RPC.

+++ protocol/schema/supervisor.capnp
@@ -29,7 +29,7 @@ interface Stdin {
 
 interface Process {
   poll @0 () -> (next :ProcessOutput);
-  signal @1 (signum :UInt8) -> ();
+  signal @1 (signum :Int32) -> ();
   kill @2 () -> ();
 }
